#172: renames relation fixture folder
