+++ test/domain/relation/explore.spec.ts
@@ -4,7 +4,7 @@ import { describe, expect, it } from 'vitest';
 import SortOptions from '^/domain/relation/definitions/SortOptions';
 import explore from '^/domain/relation/explore';
 
-import { DATABASES, REQUESTERS, VALUES } from './_fixtures';
+import { DATABASES, REQUESTERS, VALUES } from './fixtures';
 
 describe('domain/relation/explore', () =>
 {
